third_party/blink/web_tests/fast/js/negate-overflow-expected.txt

Test for overflow when negating the largest negative signed int.

If the test passes, you'll see a series of PASS messages below.

PASS: x should be 2147483648 and is.
PASS: y should be 2147483648 and is.
